requirements.txt
Django==4.2.23
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5
